+++ tienda/src/main/java/ms/tienda/repository/InventarioRepository.java
@@ -3,5 +3,5 @@ package ms.tienda.repository;
 import ms.tienda.entity.Inventario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventarioRepository extends JpaRepository<Inventario,Long> {
+public interface InventarioRepository extends JpaRepository<Inventario,Double> {
 }
